Develop

+++ src/app/services/user.services.ts
@@ -26,11 +26,11 @@ export class UserService extends SimpleRequester {
     }).toPromise();
   }
 
-  private isAuthorized() {
-    return this.doGet('/users/isAuthorized').map(res => {
+  private isAuthorized(handleError: boolean = true) {
+    return this.doGet('/users/isAuthorized', undefined, false, handleError).map(res => {
       this.globaldata.teamMember = res.headers.get('accountMember') === 'true';
       return res.json();
-    });
+    }).toPromise();
   }
 
   getPermissionsForProject(projectId: number) {
@@ -73,21 +73,14 @@ export class UserService extends SimpleRequester {
     });
   }
 
-  private checkAuth(): Promise<boolean> {
-    return this.isAuthorized().map(result => {
-      if (result) {
-        const user: User = result;
-        if (user && user.user_name) {
-          this.globaldata.currentUser = user;
-          return true;
-        }
-      } else {
-        this.globaldata.currentUser = undefined;
-        return false;
-      }
-    }, error => {
+  private async checkAuth(handleError: boolean = true): Promise<boolean> {
+    try {
+      this.globaldata.currentUser = await this.isAuthorized(handleError);
+      return true;
+    } catch (error) {
+      this.globaldata.currentUser = undefined;
       return false;
-    }).toPromise();
+    }
   }
 
   removeUser(user: User) {
@@ -97,25 +90,19 @@ export class UserService extends SimpleRequester {
     });
   }
 
-  async handleIsLogged(returnUrl) {
-    let isLogged = false;
-    await this.IsLogged().then(res => {
-      isLogged = res;
-      if (!res) {
-        this.globaldata.returnURL = returnUrl || this.router.url;
-        this.router.navigate(['/'], { queryParams: { returnUrl: returnUrl }});
-      }
-    });
-    return isLogged;
+  async redirectToLogin() {
+    if (!(await this.router.navigate(['/']))) {
+      await this.router.navigate(['/']);
+    }
   }
 
-  async IsLogged(): Promise<boolean> {
-    return this.checkAuth().catch(e => {
-      this.globaldata.returnURL = this.router.url;
-      this.router.navigate(['/'], { queryParams: { returnUrl: this.router.url }});
+  async handleIsLogged(returnUrl?: string, handleError: boolean = true): Promise<boolean> {
+    const isLogged = await this.checkAuth(handleError);
+    if (!isLogged) {
+      this.globaldata.returnURL = returnUrl || this.router.url;
       this.cookieService.remove('iio78');
-      return false;
-    });
+    }
+    return isLogged;
   }
 
   IsAdmin(): boolean {
@@ -305,7 +292,7 @@ export class UserService extends SimpleRequester {
 
   getAnyLocalPermissions() {
     if (!this.globaldata.currentUser) {
-      this.IsLogged().then(isLogged => {
+      this.handleIsLogged().then(isLogged => {
         if (isLogged) {
           return this.doGet(`/users/permissions?userId=${this.globaldata.currentUser.id}`).map(res => res.json());
         }
